Rename AlterForm container to SlidesForm
