--- src/containers/AlterForms/index.js
@@ -1,2 +1,0 @@
-export { default as SlideFormNode } from './SlideFormNode';
-export { default as SlideFormEdge } from './SlideFormEdge';
