+++ src/containers/SlidesForm/SlideFormEdge.js
