+++ src/containers/SlidesForm/SlideFormNode.js
